requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
backcall==0.1.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
colorama==0.4.1
constantly==15.1.0
cryptography==2.4.2
decorator==4.3.0
Django==1.11.15
django-redis==4.10.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.1
kombu==4.2.1
lxml==4.2.5
parso==0.3.1
pickleshare==0.7.5
prompt-toolkit==2.0.7
pycparser==2.19
Pygments==2.3.0
PyHamcrest==1.9.0
PyMySQL==0.9.2
python-Levenshtein==0.12.0
pytz==2018.7
redis==2.10.6
requests==2.20.1
selenium==3.141.0
six==1.11.0
traitlets==4.3.2
Twisted==18.9.0
urllib3==1.24.1
vine==1.1.4
wcwidth==0.1.7
zope.interface==4.6.0
